clippy fixes (#688)
* clippy fixes

* fix fmt

+++ src/device/rsp_vu_instructions.rs
@@ -844,7 +844,7 @@ pub fn vrcpl(device: &mut device::Device, opcode: u32) {
     let vte = vte(device, vt(opcode), ve(opcode) as usize);
     let input = if device.rsp.cpu.divdp {
         ((device.rsp.cpu.divin as i32) << 16)
-            | get_vpr16(device.rsp.cpu.vpr[vt(opcode) as usize], ve(opcode) as u8) as u16 as i32
+            | get_vpr16(device.rsp.cpu.vpr[vt(opcode) as usize], ve(opcode) as u8) as i32
     } else {
         get_vpr16(device.rsp.cpu.vpr[vt(opcode) as usize], ve(opcode) as u8) as i16 as i32
     };
@@ -901,7 +901,7 @@ pub fn vrsql(device: &mut device::Device, opcode: u32) {
     let vte = vte(device, vt(opcode), ve(opcode) as usize);
     let input = if device.rsp.cpu.divdp {
         ((device.rsp.cpu.divin as i32) << 16)
-            | get_vpr16(device.rsp.cpu.vpr[vt(opcode) as usize], ve(opcode) as u8) as u16 as i32
+            | get_vpr16(device.rsp.cpu.vpr[vt(opcode) as usize], ve(opcode) as u8) as i32
     } else {
         get_vpr16(device.rsp.cpu.vpr[vt(opcode) as usize], ve(opcode) as u8) as i16 as i32
     };
